fix(windows): skip the directory fsync Windows cannot perform
publishTemporaryFile fsyncs the parent directory after renaming a temporary
file into place, so the directory entry is durable. On Windows os.Open gives a
read-only handle and FlushFileBuffers needs write access, so the call failed
with "Access is denied" and took every backup, export, and quarantine
operation down with it.

Split syncDirectory: unix keeps the fsync, Windows verifies the directory and
returns. This is not a durability regression — NTFS journals the metadata
change that publishes the rename, so there is no directory entry left for user
space to force out, and bbolt and etcd resolve this the same way.

The Windows implementation keeps a Stat so a caller naming a directory that
does not exist still gets an error, which publishTemporaryFile's callers and
the maintenance edge-case test both rely on.

workstore goes from 9 failing tests to 3 on Windows. The remainder are all
one unrelated issue: assertions that files are mode 0600, which Windows
cannot express through Chmod.
